PlayerUtils deleted, all static calls changed to EMIPlayer and unused models deleted
