+++ src/main/java/com/everneth/emi/commands/AltAccountCommands.java
@@ -6,18 +6,13 @@ import co.aikar.idb.DB;
 import co.aikar.idb.DbRow;
 import com.everneth.emi.EMI;
 import com.everneth.emi.Utils;
-import com.everneth.emi.utils.PlayerUtils;
+import com.everneth.emi.models.EMIPlayer;
 import org.bukkit.Bukkit;
-import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
-import org.bukkit.command.TabCompleter;
 import org.bukkit.entity.Player;
-import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 
 import java.text.SimpleDateFormat;
 import java.util.*;
-import java.util.concurrent.CompletableFuture;
 
 /**
  * Class: AltAccountCommands
@@ -37,17 +32,17 @@ public class AltAccountCommands extends BaseCommand {
             player.sendMessage("You cannot add yourself as an alt account");
             return;
         }
-        if (!PlayerUtils.syncExists(player.getUniqueId())) {
+        if (!EMIPlayer.syncExists(player.getUniqueId())) {
             player.sendMessage(Utils.color("&7You must have a synced discord account to add an alt."));
             return;
         }
-        UUID uuid = PlayerUtils.getPlayerUUID(requestedName);
+        UUID uuid = EMIPlayer.getPlayerUUID(requestedName);
         if (uuid == null) {
             player.sendMessage("Could not find a minecraft user by that name.");
             return;
         }
 
-        DbRow dbRow = PlayerUtils.getPlayerRow(player.getName());
+        DbRow dbRow = EMIPlayer.getPlayerRow(player.getName());
         String dbUsername = dbRow.getString("player_name");
         String altUsername = dbRow.getString("alt_name");
 
@@ -55,7 +50,7 @@ public class AltAccountCommands extends BaseCommand {
         SimpleDateFormat format = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
         // the player does not have an alt already whitelisted, we want to check if the requested account is already whitelisted
         if (altUsername == null) {
-            DbRow requestedAltRow = PlayerUtils.getPlayerRow(requestedName);
+            DbRow requestedAltRow = EMIPlayer.getPlayerRow(requestedName);
 
             // if the last query returned null, the account has not been whitelisted already
             if (requestedAltRow == null) {
@@ -82,7 +77,7 @@ public class AltAccountCommands extends BaseCommand {
     @Subcommand("remove")
     @Description("Remove registered alternate account from the whitelist")
     public void onRemoveAlt(Player player) {
-        DbRow playerRow = PlayerUtils.getPlayerRow(player.getUniqueId());
+        DbRow playerRow = EMIPlayer.getPlayerRow(player.getUniqueId());
         String playerUsername = playerRow.getString("player_name");
         String altUsername = playerRow.getString("alt_name");
 
@@ -114,7 +109,7 @@ public class AltAccountCommands extends BaseCommand {
     @CommandPermission("emi.par.alt.remove")
     public void onRemoveAlt(CommandSender sender, String username)
     {
-        DbRow playerRow = PlayerUtils.getPlayerRow(username);
+        DbRow playerRow = EMIPlayer.getPlayerRow(username);
         if (playerRow == null) {
             sender.sendMessage(Utils.color("&cThere is nobody with that username."));
             return;
