+++ src/main/java/com/everneth/emi/commands/mint/projects/MintProjectCommands.java
@@ -12,7 +12,6 @@ import com.everneth.emi.models.EMIPlayer;
 import com.everneth.emi.models.mint.MintMaterial;
 import com.everneth.emi.models.mint.MintProject;
 import com.everneth.emi.models.mint.MintTask;
-import com.everneth.emi.utils.PlayerUtils;
 import org.bukkit.entity.Player;
 
 import java.util.ArrayList;
@@ -91,7 +90,7 @@ public class MintProjectCommands extends BaseCommand
             return;
         }
 
-        DbRow dbPlayerLead = PlayerUtils.getPlayerRow(lead);
+        DbRow dbPlayerLead = EMIPlayer.getPlayerRow(lead);
 
         if(dbPlayerLead == null || dbPlayerLead.isEmpty())
         {
@@ -263,7 +262,7 @@ public class MintProjectCommands extends BaseCommand
         }
 
         // Valid information is then put into the project
-        EMIPlayer emiPlayer = PlayerUtils.getEMIPlayer(player.getName());
+        EMIPlayer emiPlayer = EMIPlayer.getEMIPlayer(player.getName());
 
         project.addWorker(emiPlayer);
         player.sendMessage(Utils.color(mintProjectTag + "&aSuccessfully joined the project!"));
