+++ src/main/java/com/everneth/emi/managers/ReportManager.java
@@ -3,12 +3,11 @@ package com.everneth.emi.managers;
 import co.aikar.idb.DB;
 import co.aikar.idb.DbRow;
 import com.everneth.emi.EMI;
+import com.everneth.emi.models.EMIPlayer;
 import com.everneth.emi.models.Report;
-import com.everneth.emi.utils.PlayerUtils;
 
 import java.sql.SQLException;
 import java.text.SimpleDateFormat;
-import java.time.LocalDateTime;
 import java.util.*;
 import java.util.concurrent.CompletableFuture;
 
@@ -62,7 +61,7 @@ public final class ReportManager {
     public int messagesMissed(UUID uuid)
     {
         Report report = rm.findReportById(uuid);
-        DbRow playerRecord = PlayerUtils.getPlayerRow(uuid);
+        DbRow playerRecord = EMIPlayer.getPlayerRow(uuid);
         DbRow reportRecord = getReportRecord(uuid);
         EMI.getJda().getTextChannelById(report.getChannelId()).sendMessage("***" + playerRecord.getString("player_name") + "** has joined the game.*").queue();
         try {
@@ -79,7 +78,7 @@ public final class ReportManager {
     public List<DbRow> getMissedMessages(UUID uuid)
     {
         Report report = rm.findReportById(uuid);
-        DbRow playerRecord = PlayerUtils.getPlayerRow(uuid);
+        DbRow playerRecord = EMIPlayer.getPlayerRow(uuid);
         DbRow reportRecord = getReportRecord(uuid);
         try
         {
@@ -103,7 +102,7 @@ public final class ReportManager {
 
     public DbRow getReportRecord(UUID uuid)
     {
-        DbRow playerRow = PlayerUtils.getPlayerRow(uuid);
+        DbRow playerRow = EMIPlayer.getPlayerRow(uuid);
         try {
             CompletableFuture<DbRow> result =  DB.getFirstRowAsync("SELECT * FROM reports WHERE initiator_id = ?", playerRow.getInt("player_id"));
             return result.get();
@@ -150,7 +149,7 @@ public final class ReportManager {
     {
         Date now = new Date();
         SimpleDateFormat format = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
-        DbRow playerRow = PlayerUtils.getPlayerRow(uuid);
+        DbRow playerRow = EMIPlayer.getPlayerRow(uuid);
         DbRow reportRecord = getReportRecord(uuid);
         DB.executeUpdateAsync(
                 "UPDATE reports SET active = 0, date_closed = ? WHERE initiator_id = ?",
